requirements.txt
python-telegram-bot==20.7
requests
aiohttp
bs4
lxml
tqdm
pyrogram
tgcrypto
pymongo
telethon
yt-dlp
aria2p
flask
gunicorn
